# Simple improvement to `Rope::lines()` iterator by adding a cache