- fix: small clippy lints that don't affect public API
- nit: improve TextInfo Add|SubAssign impls
- feat: cache progression in the `Lines` iterator, making around 30% faster