@poliorcetics 2022-11-01
@pascalkuthe here are my first changes, the one I'm sure about. I'm exploring improvements to the underlying methods, but this PR alone is a 34% improvement.

### Benchmarking

I created files with line breaks at line 100, with 10k lines, 100k lines and 1 million line.

Before this PR, on my machine, it took

```
ropey-10k.txt => 10001 lines, 6.6705ms
ropey-100k.txt => 100001 lines, 67.236666ms
ropey-1m.txt => 1000001 lines, 681.34475ms
```

Now it takes:

```
ropey-10k.txt => 10001 lines, 4.445ms
ropey-100k.txt => 100001 lines, 44.790708ms
ropey-1m.txt => 1000001 lines, 454.385208ms
```

Code used for tests:

```
extern crate ropey;

use std::fs::File;
use std::io::BufReader;

use ropey::Rope;

fn main() {
    mesure("10k");
    mesure("100k");
    mesure("1m");
}

fn mesure(f: &str) {
    let text = Rope::from_reader(BufReader::new(File::open(format!("/tmp/ropey-{f}.txt")).unwrap())).unwrap();

    let start = std::time::Instant::now();
    let count = text.lines().count();
    let end = start.elapsed();
    println!("ropey-{f}.txt => {count} lines, {end:?}");
}
```

I'm not uploading the files because they're very heavy, but I can probably put them on some random hosting service if needed