@archseer 2022-11-01
Could you store a `chunk_iter: Chunks<'a>` similar to how `Bytes` and `Chars` do? It seems like LinesEnum is doing a similar job