@pascalkuthe 2022-11-01
Thanks for posting this. This version offers a nice speesbump but there is still lots of performance left on the table. This  version performs multiple binary searches (get_chunk_at_lines, RopeSlice::new). 

As @archseer mentioned we ideally want to simply iterate all chunks with the `Chunks` iterator and construct the line slices during that transversal. However then you are still calling `RopeSlice::new`. To avoid that, I have implemented a custom version of the chunks iterator that tracks the lowest common ancestor of a line. I wanted to post a PR last week but reverse iteration still base some bugs and private things got in the way. Will try to post my version as a draft later today or latest tomorrow