@cessen 2022-11-09
@poliorcetics 

I definitely appreciate the effort you've put into this!  Unfortunately @pascalkuthe's PR (#70) is really the approach I had in mind.  So I'm closing this in favor of that.  Sorry about that!