@poliorcetics 2022-11-09
Don't worry, their solution is best yes, mine was more of a stopgap :)